Copy-DbaLogin - do not overwrite files and include ObjectLevel permissions if specified (#6719)

+++ tests/Export-DbaLogin.Tests.ps1
@@ -5,7 +5,7 @@ Write-Host -Object "Running $PSCommandPath" -ForegroundColor Cyan
 Describe "$CommandName Unit Tests" -Tag 'UnitTests' {
     Context "Validate parameters" {
         [object[]]$params = (Get-Command $CommandName).Parameters.Keys | Where-Object { $_ -notin ('whatif', 'confirm') }
-        [object[]]$knownParameters = 'SqlInstance', 'SqlCredential', 'InputObject', 'Login', 'ExcludeLogin', 'Database', 'ExcludeJobs', 'ExcludeDatabase', 'ExcludePassword', 'DefaultDatabase', 'Path', 'FilePath', 'Encoding', 'NoClobber', 'Append', 'BatchSeparator', 'DestinationVersion', 'NoPrefix', 'Passthru', 'EnableException'
+        [object[]]$knownParameters = 'SqlInstance', 'SqlCredential', 'InputObject', 'Login', 'ExcludeLogin', 'Database', 'ExcludeJobs', 'ExcludeDatabase', 'ExcludePassword', 'DefaultDatabase', 'Path', 'FilePath', 'Encoding', 'NoClobber', 'Append', 'BatchSeparator', 'DestinationVersion', 'NoPrefix', 'Passthru', 'ObjectLevel', 'EnableException'
         $knownParameters += [System.Management.Automation.PSCmdlet]::CommonParameters
         It "Should only contain our specific parameters" {
             (@(Compare-Object -ReferenceObject ($knownParameters | Where-Object { $_ }) -DifferenceObject $params).Count ) | Should Be 0
@@ -17,48 +17,45 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
     BeforeAll {
         $DefaultExportPath = Get-DbatoolsConfigValue -FullName path.dbatoolsexport
         $AltExportPath = "$env:USERPROFILE\Documents"
-        try {
-            $random = Get-Random
-            $dbname1 = "dbatoolsci_exportdbalogin1$random"
-            $login1 = "dbatoolsci_exportdbalogin_login1$random"
-            $user1 = "dbatoolsci_exportdbalogin_user1$random"
+        $random = Get-Random
+        $dbname1 = "dbatoolsci_exportdbalogin1$random"
+        $login1 = "dbatoolsci_exportdbalogin_login1$random"
+        $user1 = "dbatoolsci_exportdbalogin_user1$random"
 
-            $dbname2 = "dbatoolsci_exportdbalogin2$random"
-            $login2 = "dbatoolsci_exportdbalogin_login2$random"
-            $user2 = "dbatoolsci_exportdbalogin_user2$random"
+        $dbname2 = "dbatoolsci_exportdbalogin2$random"
+        $login2 = "dbatoolsci_exportdbalogin_login2$random"
+        $user2 = "dbatoolsci_exportdbalogin_user2$random"
 
-            $server = Connect-DbaInstance -SqlInstance $script:instance2
-            $null = $server.Query("CREATE DATABASE [$dbname1]")
-            $null = $server.Query("CREATE LOGIN [$login1] WITH PASSWORD = 'GoodPass1234!'")
-            $server.Databases[$dbname1].ExecuteNonQuery("CREATE USER [$user1] FOR LOGIN [$login1]")
+        $server = Connect-DbaInstance -SqlInstance $script:instance2
+        $null = $server.Query("CREATE DATABASE [$dbname1]")
+        $null = $server.Query("CREATE LOGIN [$login1] WITH PASSWORD = 'GoodPass1234!'")
+        $server.Databases[$dbname1].ExecuteNonQuery("CREATE USER [$user1] FOR LOGIN [$login1]")
 
-            $null = $server.Query("CREATE DATABASE [$dbname2]")
-            $null = $server.Query("CREATE LOGIN [$login2] WITH PASSWORD = 'GoodPass1234!'")
-            $null = $server.Query("ALTER LOGIN [$login2] DISABLE")
-            $null = $server.Query("DENY CONNECT SQL TO [$login2]")
+        $null = $server.Query("CREATE DATABASE [$dbname2]")
+        $null = $server.Query("CREATE LOGIN [$login2] WITH PASSWORD = 'GoodPass1234!'")
+        $null = $server.Query("ALTER LOGIN [$login2] DISABLE")
+        $null = $server.Query("DENY CONNECT SQL TO [$login2]")
 
-            if ($server.VersionMajor -lt 11) {
-                $null = $server.Query("EXEC sys.sp_addsrvrolemember @rolename=N'dbcreator', @loginame=N'$login2'")
-            } else {
-                $null = $server.Query("ALTER SERVER ROLE [dbcreator] ADD MEMBER [$login2]")
-            }
-            $null = $server.Query("GRANT SELECT ON sys.databases TO [$login2] WITH GRANT OPTION")
-            $server.Databases[$dbname2].ExecuteNonQuery("CREATE USER [$user2] FOR LOGIN [$login2]")
-        } catch {
-            $_
+        if ($server.VersionMajor -lt 11) {
+            $null = $server.Query("EXEC sys.sp_addsrvrolemember @rolename=N'dbcreator', @loginame=N'$login2'")
+        } else {
+            $null = $server.Query("ALTER SERVER ROLE [dbcreator] ADD MEMBER [$login2]")
         }
+        $server.Databases[$dbname2].ExecuteNonQuery("CREATE USER [$user2] FOR LOGIN [$login2]")
+        $server.Databases[$dbname2].ExecuteNonQuery("GRANT SELECT ON sys.tables TO [$user2] WITH GRANT OPTION")
+
     }
     AfterAll {
-        try {
-            Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname1 -Confirm:$false
-            Remove-DbaLogin -SqlInstance $script:instance2 -Login $login1 -Confirm:$false
+        Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname1 -Confirm:$false
+        Remove-DbaLogin -SqlInstance $script:instance2 -Login $login1 -Confirm:$false
 
-            Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname2 -Confirm:$false
-            Remove-DbaLogin -SqlInstance $script:instance2 -Login $login2 -Confirm:$false
-        } catch { }
+        Remove-DbaDatabase -SqlInstance $script:instance2 -Database $dbname2 -Confirm:$false
+        Remove-DbaLogin -SqlInstance $script:instance2 -Login $login2 -Confirm:$false
         $timenow = (Get-Date -uformat "%m%d%Y%H")
         $ExportedCredential = Get-ChildItem $DefaultExportPath, $AltExportPath | Where-Object { $_.Name -match "$timenow\d{4}-login.sql|Dbatoolsci_login_CustomFile.sql" }
-        $null = Remove-Item -Path $($ExportedCredential.FullName) -ErrorAction SilentlyContinue
+        if ($ExportedCredential) {
+            $null = Remove-Item -Path $($ExportedCredential.FullName) -ErrorAction SilentlyContinue
+        }
     }
 
     Context "Executes with Exclude Parameters" {
@@ -75,10 +72,14 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
             $results | Should Not Match 'Job'
         }
         It "Should exclude Go when exporting" {
-            $file = Export-DbaLogin -SqlInstance $script:instance2 -ExcludeDatabase -BatchSeparator '' -WarningAction SilentlyContinue
+            $file = Export-DbaLogin -SqlInstance $script:instance2 -BatchSeparator '' -ObjectLevel -WarningAction SilentlyContinue
             $results = Get-Content -Path $file -Raw
             $allfiles += $file.FullName
-            $results | Should Not Match 'Go'
+            $results | Should Not Match 'GO'
+            $results | Should Match "GRANT SELECT ON OBJECT::\[sys\]\.\[tables\] TO \[$user2\] WITH GRANT OPTION"
+            $results | Should Match "CREATE USER \[$user2\] FOR LOGIN \[$login2\]"
+            $results | Should Match "IF NOT EXISTS"
+            $results | Should Match "USE \[$dbname2\]"
         }
         It "Should exclude a specific login" {
             $file = Export-DbaLogin -SqlInstance $script:instance2 -ExcludeLogin $login1 -WarningAction SilentlyContinue
@@ -100,6 +101,14 @@ Describe "$CommandName Integration Tests" -Tags "IntegrationTests" {
             $allfiles += $file.FullName
             $results | Should Not Match "$login2|$dbname2"
             $results | Should Match "$login1|$dbname1"
+        }
+        It "Should Export with object level permissions" {
+            $results = Export-DbaLogin -SqlInstance $script:instance2 -Login $login2 -ObjectLevel -PassThru -WarningAction SilentlyContinue
+            $results | Should Not Match "$login1|$dbname1"
+            $results | Should Match "GRANT SELECT ON OBJECT::\[sys\]\.\[tables\] TO \[$user2\] WITH GRANT OPTION"
+            $results | Should Match "CREATE USER \[$user2\] FOR LOGIN \[$login2\]"
+            $results | Should Match "IF NOT EXISTS"
+            $results | Should Match "USE \[$dbname2\]"
         }
         foreach ($version in $((Get-Command $CommandName).Parameters.DestinationVersion.attributes.validvalues)) {
             It "Should Export for the SQLVersion $version" {
